playtime

+++ src/code/game/g_local.h
@@ -190,7 +190,7 @@ public:
    qboolean          force_entnum;
    int               spawn_entnum;
 
-   float             playtime;
+   int               playtime;
 
    // List of valid player models loaded from players global scriptfile
    Container<str>    ValidPlayerModels;
@@ -262,7 +262,7 @@ public:
    qboolean    training;
 
    Entity      *defaultcamera;
-   float       playtime;
+   int         playtime;
 
    level_locals_t();
    virtual void   Archive(Archiver &arc)   override;
